gateway funcionando, entre otras cosas. Se puede consultar y crear pruebas, empleados e interesados

+++ empleados/src/main/java/utn/frc/backend/empleados/dto/EmpleadoDTO.java
@@ -11,5 +11,5 @@ public class EmpleadoDTO {
     private long legajo;
     private String nombre;
     private String apellido;
-    private int telefonoContacto;
+    private long telefonoContacto;
 }
