+++ pruebas/src/main/java/utn/frc/backend/pruebas/dto/EmpleadoDTO.java
@@ -7,5 +7,5 @@ public class EmpleadoDTO {
     private long legajo;
     private String nombre;
     private String apellido;
-    private int telefonoContacto;
+    private long telefonoContacto;
 }
